requirements.txt
opencc-python-reimplemented
pyewts
rdflib
polib